requirements.txt
streamlit
wikipedia
beautifulsoup4
requests
pandas
google-generativeai

